Move test to preferred directory
Add a new test location for the submission validation test
Update the paths and objects needed.
